Change log dependency to 0.4.x
Log dependency introduces several naming changes in API. Besides that
the most significant change for us is lifetime change for record location
strings from 'static to 'a. This requires more changes in slog and then in
slog-async.

log -> slog message transition is implemented and working (alongisde with
slog and slog-async changes).

slog -> log message transition is not implemented yet properly (it is
compiling though).