Avoid using log's private API to construct records
Previously this API was required to work around rust-lang/log#95

Thanks to the inclusion of a `RecordBuilder` (PR rust-lang/log#211)
it is no longer required.

This fixes slog-rs/slog#309
This supersedes PR #20

I have added a comment about supporting the log/kv_unstable feature in the future.
Not supporting this feature doesn't break backwards compat, because the
old usage of the private API didn't support it either. Therefore, this
new code doesn't break or remove anything that was already existing.
However,from a feature-completeness standpoint it would be nice to have
this in the future. (see comment in code)