Bump claimed MSRV to 1.39
This is required for the log/kv_unstable feature.
This doesn't break 1.38 compatibility in any way shape or form.
It mearly changes the documentation and the official guarentees.

If you want, you can still use Rust 1.38. However, you will need to disable the
log/kv_unstable feature.

The advantage of this change is that it makes Github Actions testing easier.
e only need to test a single MSRV, instead of a different one for
each feature combination :)